requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.0
